fix all warnings

+++ scc/RSA.cpp
@@ -7,8 +7,8 @@
 
 scc::RSA::RSA(scb::Bytes const &modulus, scb::Bytes const &exponent) {
     rsa_ = RSA_new();
-    modulus_ = BN_bin2bn(modulus.data(), modulus.size(), NULL);
-    exponent_ = BN_bin2bn(exponent.data(), exponent.size(), NULL);
+    modulus_ = BN_bin2bn(modulus.data(), static_cast<int>(modulus.size()), NULL);
+    exponent_ = BN_bin2bn(exponent.data(), static_cast<int>(exponent.size()), NULL);
 
     int result = RSA_set0_key(rsa_, modulus_, exponent_, NULL);
 
@@ -24,7 +24,7 @@ scc::RSA::~RSA() {
 
 scb::Bytes scc::RSA::transorm(scb::Bytes const &bytes) const {
     scb::Bytes result(BN_num_bits(modulus_) / 8);
-    if (RSA_public_decrypt(bytes.size(), bytes.data(), result.data(), rsa_, RSA_NO_PADDING) == -1) {
+    if (RSA_public_decrypt(static_cast<int>(bytes.size()), bytes.data(), result.data(), rsa_, RSA_NO_PADDING) == -1) {
         throw Exception(ERR_get_error());
     }
     return result;
